refactor(#56): BuildingDataSource 위치 변경
