+++ app/src/main/java/com/ssafy/tott/data/repository/BuildingRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.ssafy.tott.data.repository
 
-import com.ssafy.tott.data.datasource.BuildingDataSource
+import com.ssafy.tott.data.datasource.remote.BuildingRemoteDataSource
 import com.ssafy.tott.data.model.request.SearchFilterRequest.Companion.toData
 import com.ssafy.tott.data.model.response.HouseDetailResponse
 import com.ssafy.tott.domain.model.Building
@@ -10,16 +10,16 @@ import com.ssafy.tott.domain.repository.BuildingRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
-class BuildingRepositoryImpl(private val buildingDataSource: BuildingDataSource) :
+class BuildingRepositoryImpl(private val buildingRemoteDataSource: BuildingRemoteDataSource) :
     BuildingRepository {
     override fun getBuildings(searchFilter: SearchFilter): Flow<Result<List<Building>>> {
-        return buildingDataSource.searchBuilding(searchFilter.toData()).map { result ->
+        return buildingRemoteDataSource.searchBuilding(searchFilter.toData()).map { result ->
             result.map { response -> response.buildingListResponse.map { it.toDomain() } }
         }
     }
 
     override fun getRecentBuildingDetails(ids: List<Int>): Flow<Result<List<HouseSaleArticle>>> {
-        return buildingDataSource.getRecentHouseArticles(ids).map { result ->
+        return buildingRemoteDataSource.getRecentHouseArticles(ids).map { result ->
             result.map {
                 it?.houseDetailResponseList?.map(HouseDetailResponse::toDomain) ?: emptyList()
             }
@@ -27,7 +27,7 @@ class BuildingRepositoryImpl(private val buildingDataSource: BuildingDataSource)
     }
 
     override fun getWishList(): Flow<Result<List<HouseSaleArticle>>> {
-        return buildingDataSource.getWishList().map { result ->
+        return buildingRemoteDataSource.getWishList().map { result ->
             result.map {
                 it?.houseDetailResponseList?.map(HouseDetailResponse::toDomain) ?: emptyList()
             }
